# server/api/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      timestamp:
        description: 响应时间（RFC3339 格式）
        type: string
      trace_id:
        description: 链路追踪 ID（用于日志关联和分布式追踪）
        type: string
    type: object
  github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      timestamp:
        description: 响应时间（RFC3339 格式）
        type: string
      trace_id:
        description: 链路追踪 ID（用于日志关联和分布式追踪）
        type: string
    type: object
...
  internal_transport_http_handlers.batchUpdateRequest:
    properties:
      settings:
        items:
          $ref: '#/definitions/internal_transport_http_handlers.settingItem'
        type: array
    required:
    - settings
    type: object
  internal_transport_http_handlers.settingItem:
    properties:
      key:
        type: string
      value:
        items:
          type: integer
        type: array
    required:
    - key
    - value
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.github.com/shenfay/go-react-admin
  description: 生产级 DDD 脚手架项目的 API 文档，包含用户认证、事件驱动等核心功能。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go DDD Scaffold API
  version: "1.0"
paths:
...
  /settings:
    get:
      parameters:
      - description: 设置分类
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 设置列表
          schema:
            $ref: '#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取系统设置列表
      tags:
      - Settings
    put:
      consumes:
      - application/json
      parameters:
      - description: 批量更新数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_handlers.batchUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 批量更新系统设置
      tags:
      - Settings
  /settings/{key}:
    get:
      parameters:
      - description: 设置项Key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 设置详情
          schema:
            $ref: '#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse'
        "404":
          description: 设置项不存在
          schema:
            $ref: '#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取单个系统设置
      tags:
      - Settings
...
securityDefinitions:
  BearerAuth:
    description: 使用 JWT Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"